% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neff.R
\name{calc_neff}
\alias{calc_neff}
\title{Compute Peto's Effective Sample Size from a survfit Object}
\usage{
calc_neff(fit_km)
}
\arguments{
\item{fit_km}{A \code{survfit} object (stratified or unstratified).}
}
\value{
Data frame: time, n_risk, n_event, surv, var_s, neff, arm.
}
\description{
Compute Peto's Effective Sample Size from a survfit Object
}
